# swagger.yaml
openapi: 3.0.3
info:
  title: Swagger API Course FE
  description: |-
    This is a list api of course FE
  contact:
    email: vanthang.thophu@gmail.com
  version: 1.0.11
servers:
  - url: https://backend-nextjs.onrender.com/
  - url: http://localhost:3002
...
paths:
...
  /api/chat/suggestions:
    get:
      tags:
        - chat
      summary: Lấy danh sách câu hỏi gợi ý
      description: Trả về danh sách các câu hỏi mẫu để người dùng có thể bắt đầu hội thoại
      operationId: getSuggestedQuestions
      responses:
        "200":
          description: Lấy gợi ý thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "Success"
                  message:
                    type: string
                    example: "Lấy gợi ý câu hỏi thành công"
                  typeError:
                    type: string
                    example: ""
                  data:
                    type: object
                    properties:
                      questions:
                        type: array
                        items:
                          type: string
                        example:
                          - "Tìm sản phẩm dưới 500 nghìn"
                          - "Sản phẩm bán chạy nhất là gì?"
                          - "Có sản phẩm nào đang giảm giá không?"
                          - "Gợi ý sản phẩm cho tôi"
        "500":
          description: INTERNAL_SERVER_ERROR
  /api/chat/search:
    post:
      tags:
        - chat
      summary: Tìm kiếm sản phẩm bằng ngôn ngữ tự nhiên
      description: Tìm kiếm sản phẩm dựa trên câu truy vấn ngôn ngữ tự nhiên (không qua AI)
      operationId: searchProducts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: Câu truy vấn tìm kiếm
                  example: "giày dưới 1 triệu màu đỏ"
              required:
                - query
      responses:
        "200":
          description: Tìm kiếm thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "Success"
                  message:
                    type: string
                    example: "Tìm kiếm thành công"
                  typeError:
                    type: string
                    example: ""
                  data:
                    type: object
                    properties:
                      products:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            name:
                              type: string
                            slug:
                              type: string
                            image:
                              type: string
                            price:
                              type: number
                            discount:
                              type: number
                            actualPrice:
                              type: number
                            type:
                              type: string
                            countInStock:
                              type: number
                      filters:
                        type: object
                      total:
                        type: number
        "400":
          description: INVALID - Query không hợp lệ
        "500":
          description: INTERNAL_SERVER_ERROR

  # ==================== AUTH API ====================
...
  /api/notifications/delete-many:
    delete:
      tags:
        - notifications
      summary: Delete many notifications
      description: Delete multiple notifications by IDs
      operationId: deleteManyNotifications
      requestBody:
        description: Array of notification IDs to delete
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                notificationIds:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "400":
          description: INVALID
      security:
        - bearerAuth: []
  /api/notifications/{notificationId}:
    delete:
      tags:
        - notifications
      summary: Delete notification by id
      description: Delete notification by id
      operationId: deleteNotification
      parameters:
        - name: productId
          in: path
          description: ID of notification that needs to be fetched
          required: true
          schema:
            type: string
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Notification"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Notification"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    Notification:
      type: object
      properties:
        _id:
          type: string
          format: uuid
        recipientIds:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
                format: uuid
              isRead:
                type: boolean
        referenceId:
          type: string
          format: uuid
        body:
          type: string
        title:
          type: string
        context:
          type: string
      required:
        - context
        - title
        - body
      xml:
        name: Notification
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT